fix: refactor and cleanup
