+++ scripts/modules/ui.js
@@ -9,8 +9,6 @@ import boxen from 'boxen';
 import ora from 'ora';
 import Table from 'cli-table3';
 import gradient from 'gradient-string';
-import { fileURLToPath } from 'url'; 
-import path, { dirname, join as pathJoin } from 'path'; 
 import fs from 'fs';
 import {
 	log,
@@ -21,6 +19,7 @@ import {
 } from './utils.js';
 import { findNextTask, analyzeTaskComplexity } from './task-manager.js';
 import { getProjectName, getDefaultSubtasks } from './config-manager.js';
+import { getTaskMasterVersion } from '../../src/utils/getVersion.js';
 
 // Create a color gradient for the banner
 const coolGradient = gradient(['#00b4d8', '#0077b6', '#03045e']);
@@ -47,25 +46,7 @@ function displayBanner() {
 	);
 
 	// Read version directly from package.json
-	let version = 'unknown'; // Initialize with a default
-	try {
-		// Get the directory of the current module (ui.js)
-		const currentModuleFilename = fileURLToPath(import.meta.url);
-		const currentModuleDirname = dirname(currentModuleFilename);
-		// Construct the path to package.json relative to ui.js (../../package.json)
-		const packageJsonPath = pathJoin(currentModuleDirname, '..', '..', 'package.json');
-
-		if (fs.existsSync(packageJsonPath)) {
-			const packageJsonContent = fs.readFileSync(packageJsonPath, 'utf8');
-			const packageJson = JSON.parse(packageJsonContent);
-			version = packageJson.version;
-		} else {
-			log('warn', `Own package.json not found at expected path: ${packageJsonPath}`);
-		}
-	} catch (error) {
-		// Silently fall back to default version
-		log('warn', 'Could not read own package.json for version info.', error);
-	}
+	const version = getTaskMasterVersion();
 
 	console.log(
 		boxen(
@@ -946,7 +927,7 @@ async function displayNextTask(tasksPath) {
 						}
 						return chalk.red(`${nextTask.id}.${depId} (Not found)`);
 					}
-					return depId; // Assume it's a top-level task ID if not a number < 100
+					return depId;
 				});
 
 				// Join the formatted dependencies directly instead of passing to formatDependenciesWithStatus again
@@ -1238,7 +1219,6 @@ async function displayTaskById(tasksPath, taskId, statusFilter = null) {
 			const statusColor = statusColorMap[st.status || 'pending'] || chalk.white;
 			let subtaskDeps = 'None';
 			if (st.dependencies && st.dependencies.length > 0) {
-				// Format dependencies with correct notation
 				const formattedDeps = st.dependencies.map((depId) => {
 					// Use the original, unfiltered list for dependency status lookup
 					const sourceListForDeps = originalSubtasks || task.subtasks;
